Moved dist/ to input/dist_template
Originally the dist/ folder was created to house the formal output of the make process,
however with 'make rpm' the need for a specifically named output directory ( %{name}-%{version} ) was
created. This moves dist/ into input/dist_template, so we can use it as a skeleton directory. The
'make dist' now copies input/dist_template to ${name}-%{version}, keeping all subfolders and .gitignore files.