Switch from curly braces to parens in Makefile
I haven't seen this before but it looks like curly
braces causes variables to be treated as simply
expanded variables even if they are declared
with a " = " instead of " := ".  We need certain
variables to be recursively expanded vars so
might as well switch to parens everywhere to
make things consistent and less error-prone.