Fix tarball deps and tar.
Recursively expanded vars used as target deps are expanded
early when make is constructing the dep graph.  Thus, $(TAR_DEPS)
will be empty when evaluated.  For now use the $(DIST) as a quick
fix.

We need to cleanup all the deps as we pretty much ignore this
in favor of targets that will always be ran.