Mapped CCI-001462 to service_auditd_enabled
CCI-001462 requires the "capability" to audit ALL user session information. In reality this is implimented by turning on key logging, however since we only have to provide the "capability" vs the "implimentation" we're calling this met by turning on audit.

It should be noted that this "capability" is actually met just by installing the auditd package, however that doesn't tie very nicely with anything in our existing content. At some point we should create a rule to check that audit is actually installed (not configured, but installed) and tie CCI-001462 to that rule.