Mapped CCI-001115 to enable_ip6tables, enable_iptables, auditd_configure_rules and enable_auditd_service
This CCI requires that we can deny network traffic and audit potentially malicious users/programs.

Network traffic denial is done by blocking all but needed ports (which is done in the default firewall rules once iptables is enabled), and we treat ALL users (potentially malicious or not) the same in the auditing system once it's enabled and configured.