Mapped CCI-001232 to security_patches_up_to_date.
CCI-001232 requires automatic installation of software updates. RHEL6 can be configured to do this using yum, but it is not recommended in production systems. Guidance should be added to the software updates section of the guide to provide guidance on updating best practices. A ticket has been created in Trac for this.