added new transform to reformat our XCCDF into STIG-like structure
This is incomplete.  For now, it merely adds a CCI to those items
which do not yet have one.

In the future, this transform will strip out all Rules which are not part of the
STIG profile, and also restructure the entire document as a list
of Groups, each of which contains one Rule.