Request for discussion: Removed AU-2 mappings from system/logging.xml
Only the audit subsystem should be considered an
authoritative source of audit information. Removing
mappings of AU-2 to syslog services for this reason.
Note this isn't to say setting up syslog isn't a good,
prudent thing to do -- but should not be considered
authoritative and thus should not be considered to meet
AU-2* requirements.