New Profile: C2S
Adding new profile, C2S in support of a U.S. Intelligence Community baseline. This baseline is specifically derived from CIS' RHEL6 control list. Those who are interested in CIS baseline should be able to follow C2S; will work with the CIS guys in the future to get permission to directly call this "CIS", though that's a lower priority when measured against completing the XCCDF/OVAL components.

As indicated in the profile description, this baseline is NOT complete -- however wanted to get progress out on the list. At this point the profile will scan the system without errors, though next steps include:
- Adding ~64 additional OVAL checks.
- Ensuring refine values match baseline standards (currently, inheriting default values from XCCDF)
- Remediation scripts