exclude proc/selinux files from world_writeable_files
The following commit in openscap causes proc to be considered a 'local'
filesystem.
---
commit a9d98d697d85cb7f26c615543276e042c9749af0
Author: Tomas Heinrich <theinric@redhat.com>
Date:   Tue Nov 12 11:01:21 2013 +0100

    OVAL: change the algo used to distinguish 'local' and 'remote' fs

    With 'local', fts now descends into /proc and /sys, among others,
    which were previously excluded. This can cause performance issues.
---

The 'world_writeable_files' test will fail if a world-writeable
regular file is found. So we must mask out /proc files with a filter.
I did a `find / -perm -002 -type f` and also found some 'special' files
in /selinux that are world-writeable and type f, so a filter has been
added for those too.