[bugfix] [RHEL/6/input/checks/templates] Provide create_umask_checks.py template for generating (super-compliance) OVAL umask implementation.
The current OVAL umask checks (e.g. accounts_umask_etc_profile.xml) aren't
super-compliant as they PASS only when system configured (e.g. /etc/profile)
umask exactly matches the requirement.

To be super-compliant the umask OVAL check should PASS also in cases when
system in question is configured to have more strict umask in particular file.
For example, taking /etc/profile as example, the requirement umask to be '002',
then also umasks of the form e.g. '007' or '022' should PASS this requirement
(since 002 in binary form is subset of 007 or 022 in binary form).

The new check is implemented as follows - we first create octal representation
of the value retrieved from the particular config file, then create octal
representation of required value provided in external OVAL variable. At the
end compare the binary representations of these two octal representations via
the 'bitwise and' OVAL operator.

Example:

  Requirement:   002 => 0*64 + 0*8 + 2*1 = 2
  Config value:  007 => 0*64 + 0*8 + 7*1 = 7

Then perform bitwise AND on binary representations of the two:

  R  (in binary):       000 000 010
  CV (in binary):       000 000 111
  ---------------------------------
  bitwise AND           000 000 010

Since bitwise AND result is the same as requirement (in binary) the result
PASSED & corresponding umask is allowed. Computational example of umask
that wouldn't pass this check is left as exercise for the reader.