[BugFix] [RHEL/7] [Fedora] Add a note into the introductory XCCDF section for the auditd service that on RHEL-7 and Fedora auditd service can be configured to use two tools (augenrules or auditctl) to read audit.rules configuration.
The note is expected to differ for the two products RHEL-7 and Fedora
because while on RHEL-7 the default tool (default configuration) is
augenrules, in Fedora the default tool (default configuration) is
auditctl. So modify the appropriate part of the description for Fedora
the default tool to be described first.