[BugFix] [RHEL/6] [RHEL/7] Don't include the 'test' profile into the final benchmark by default, only upon request
This change modifies the Makefile's for RHEL/6 and RHEL/7 products
in order to the 'test' profile not to be included into the produced
RHEL-6 and RHEL-7 benchmark by default, since it is profile dedicated
for testing and not intended for production use. Both Makefiles have
been modified to include the 'test' profile only in moment, it's
inclusion was request (e.g. when evaluation 'eval-test' target).

This fixes the following downstream bug:
  [1] https://bugzilla.redhat.com/show_bug.cgi?id=1153734

While on the Makefiles, this commit fixes also the following bugs
(in both Makefiles):
* substitute couple of 'cd output;' calls with 'cd $(OUT);' equivalents,
* the 'eval-common' target has been previously missing the 'content'
  prerequisite (so it didn't work when run without first running 'make'
  target for that product),
* also, as written in the oscap(8) manual page "oscap" when run with
  --oval-results option always places the generated results file into CWD,
  which previously led into polluting CWD with OVAL results file. Therefore
  update the 'eval common' target to first "cd" into $(OUT) directory,
  and call "oscap eval" only then.

Testing report:
---------------
The change has been tested ("make", "make eval-test", "make eval-common")
on the following systems:
* RHEL-6,
* RHEL-7, and
* Fedora 21

and it has been verified it works as expected.