[Enhancement] [RHEL/7] Add new RHEL-7 remediations for the following audit DAC 'chown' group rules: * audit_rules_dac_modification_chown, * audit_rules_dac_modification_fchown, * audit_rules_dac_modification_fchownat, and * audit_rules_dac_modification_lchown.
Note: It is intended the four remediation fixes are identical (one
script fixes the settings for all four of the system calls). Since
it's not good to use symbolic links, we use multiple copies.

Testing report:
---------------
The proposed change has been tested on RHEL-7 system and seems to be
working fine.