Implemented unselect-empty-xccdf-groups.py
The script unselects XCCDF groups that have no selected XCCDF rules in
every profile of given XCCDF Benchmark. This is useful to make HTML
guides of very small profiles much shorter and avoid lengthy
descriptions of XCCDF Groups that are not even used.

The script is only used in Fedora content build for now. I will get it
right for Fedora first, then apply it to all content.