[Enhancement] [RHEL/7] New RHEL-7 specific remediation for aide_build_database rule
It's RHEL-7 specific because the aide executable path on different
products would be different. In the future we might add more generic
version checking appropriate product version, and based on that
version change the path. But that would be future enhancement.

Testing report:
---------------
Verified the proposed change works fine on RHEL-7 system.