[Enhancement] [RHEL/7] Switch on use of 'install_hids' rule
There isn't a default recommended IDS system / RPM package in
RHEL-7 (besides aide and IDS checking via rpm), therefore it's
hard to write an OVAL check to test all of the possible IDS combinations
that can happen. In the future the OVAL check might be implemented
(for now the rule will be just displayed in the guide and result would
be 'notchecked').