[Enhancement] [RHEL/7] New RHEL-7 remediation for 'rsyslog_files_permissions' rule
NOTE: The machine in question needs to be rebooted before
the system is put into the true compliance (this is due to
handling of "/var/log/boot.log" case), therefore IT IS EXPECTED
this remediation when run will return ERROR when attempted in
openscap. To truly verify if the system is corrected, the system
needs to be rebooted, and the scan for the 'rsyslog_files_permissions'
rule repeated.

Testing report:
---------------
Verified on RHEL-7 system the proposed change works fine (after reboot
the 'rsyslog_files_permissions' rule passes).