[BugFix] [RHEL/6] [RHEL/7] Fix remediations for: * 'file_permissions_binary_dirs', and * 'file_ownership_binary_dirs'
to honour also '/usr/libexec' location when performing the corrections.
Also drop the RHEL-6 specific remediation for 'file_permissions_binary_dirs'
(it was identical to RHEL-7 version), and move RHEL-7 version of this
remediation to be shared/ one (updating platform tag appropriately).

Testing report:
Verified manually on both systems:
* RHEL-6, and
* RHEL-7

the updated remediation scripts work fine on both of the products.