[BugFix] [RHEL/5] [RHEL/6] [RHEL/7] [Fedora] Update existing OVALs for 'audit_rules_time_clock_settime' rule in order they to check for presence of '-F a0=0x0' argument too.
This is the OVAL fix part of:
  [1] https://github.com/OpenSCAP/scap-security-guide/issues/818

Testing report:
--------------
Verified manually on both RHEL-6 and RHEL-7 systems the proposed
change works fine (either -F a0=0 or -F a0=0x0 forms are required
in order the updated OVALs for RHEL-5+ to pass). Also verified
both forms ("-F key=time-change" and "-k time-change") of specifying
audit key for this rule are supported.