[Enhancement] [RHEL/5] [RHEL/6] [RHEL/7] Provide links to remote (offical Red Hat RHSA / CVE OVAL) for 'security_patches_up_to_date' rule to be executed when system is scan against this rule.
Since OpenSCAP doesn't support 'https://' protocol yet for retrieving
the remote content (see https://fedorahosted.org/openscap/ticket/213 for
further details), use 'http://' as the protocol for now when retrieving
that content (to be enhanced in the future).

Import note: To be able to scan the system against this OVAL the
"--fetch-remote-resources" option of the oscap tool is required to
be specified (similarly for SCAP Workbench 'Fetch Remote Resources'
checkbox needs to be checked in).

Testing report:
---------------
The proposed change has been tested on recent versions of RHEL-6
and RHEL-7 systems (for both XCCDF and DS format), and AFAICT
from testing they are working fine (in both cases).