Add Chromium, Firefox and JRE to "make install"
This allows displaying guides for those benchmarks on OpenSCAP portal.
The guides for the portal are built by Jenkins that runs make install
and then they are uploaded to web. It means when we want to show these
guides on the portal, we need to add them to make install target.