[BugFix] [RHEL/6] [RHEL/7] Fix the RHEL-6 & RHEL-7 OVALs for 'kernel_module_bluetooth_disabled' rule
Per testing result in:
  https://bugzilla.redhat.com/show_bug.cgi?id=1223865#c0

the only way how to disable both:
* prevent bluetooth service from starting, and simultaneously
* prevent kernel bluetooth module from loading (explicitly e.g.
  via modprobe)

on RHEL-6 & RHEL-7 is to place 'install bluetooth /bin/true'
statement in some of the various modprobe config files.

Right now the shared/ version of the OVAL is too liberal (it passes
when either of 'install net-pf-31 /bin/true' or 'install bluetooth /bin/true'
statement is configured in some of the modprobe config files. Thus
change this to require explicitly 'bluetooth' to be listed there.

Also, since we have new kernel module disabled template for RHEL-7,
the shared/ version is not applicable for RHEL-7 anymore (there are
three more locations where this differs from RHEL-6 version).

Therefore delete the original /shared version, and replace the OVAL
with two separate OVALs (one for RHEL-6, another one for RHEL-7)
using templates for creation of both of them.

Fixes:
  https://bugzilla.redhat.com/show_bug.cgi?id=1223865