Create xccdf-unlinked-unresolved.xml as a separate target
Note that I have changed the name of the file in the output/ directory.
The type of the file (xccdf) goes first, then comes the state
(unlinked-unresolved). The product name is no longer part of the
filename.

The motivation behind this is to enable easy bash completion when
debugging. I'll slowly amend all the files in the directory.