[Enhancement][RHEL/7] Move GNOME XCCDF content into its own gnome.xml XCCDF file
As more and more GNOME and GDM checks are added to the STIG and SSG, it doesn't
make sense to have GNOME/GDM checks in a bunch of disparate files. This centralizes
much of the GNOME content into a single file except for cases where GNOME and other
rules will implement the same shared variable.

- Move majority of GNOME XCCDF content (excluding banners) into gnome.xml.
- Preps for addition GNOME/GDM STIG settings
- Part of #481