[Bugfix] Revisit OVAL for "accounts_max_concurrent_login_sessions" rule #1329
This improves the logic of the maxlogins configuration file
contents.

If limits.d/*.conf does not specify maxlogins at all, but
limits.conf specifies maxlogins appropriately, then PASS

If limits.d/*.conf specifies maxlogins correctly, then PASS
regardless of the contents of limits.conf

It is still not 100% correct.  The remaining issue is:

1)  If limits.d/*.conf specifies a correct value and an incorrect
    value, this version will always FAIL.  It should pass if the
    last value is appropriate, regardless of previous values.

Signed-off-by: T.O. Radzy Radzykewycz <radzy@windriver.com>