combine-ovals: Add check for overriding OVAL 5.11 by 5.10
We want to allow overriding of shared OVAL by product OVAL files.
But we can also have shared OVAL in version 5.11 and product OVAL
in version 5.10. In such situation it is not clear which version
should be used. So this change will disable this option. If
contributor would need to override shared oval 5.11 by
product oval 5.10, he/she can create file with OVAL 5.10
in OVAL 5.11 directory