Enable rules returning 'notapplicable' in standard profile
Rules 'mount_option_dev_shm_nodev' and 'mount_option_dev_shm_nosuid'
are marked as 'machine' only, so they will return 'notapplicable'
anyway. Enable them so that machines can take benefit of it.

Enable 'security_patches_up_to_date' rule in standard profile, it
removed in #1303 because it doesn't make sense for containers,
but it does for RHEL7 bare-metal and virtual machines.
So adding it back, and marking the rule as 'machine' only.