RHBZ #1403905: Fix rules for removable media properties
The OVAL definitions assumed that if a machine has a CD drive,
the CD drive must have a configuration entry in /etc/fstab.
When it isn't there at all, the rules resulted in fail.
This commit fixes those false positives by allowing this situation
in mount_option_.*_removable_partition rules.