Improve rule for sshd_use_approved_macs
Use an XCCDF value to enable the custiomization and
avoid bugs similar to RHBZ #1413494 in future.
The rule allowed only exact list of the MACs, in a required order.
If user removes some of the MACs from /etc/ssh/sshd_config.
the configuration still fulfills the requirement,
but a fail was reported, because the regular expression required
all of them listed.