Drop OVAL checks whose extend_definition refs don't exist
Building OVAL 5.10 only content was breaking due to non-existing checks
referenced in extend_definition elements.

I considered this situation to be the same as when an XCCDF Rule
references an OVAL check witch doesn't exist, an incomplete check
is as useful as a non-existing check.

So the OVAL check with the broken extend_definition is removed and the
XCCDF Rule referencing it will have the check-content element removed
in the following steps.