Make rsyslog_remote_loghost scapval compliant
NIST scapval validation tool complains that a variable passed to
rsyslog_remote_loghost OVAL check from the XCCDF Rule doesn't have
the correct type according to the SCAP specifications.

This happens because we don't use the received variable in the check,
thus its type is not defined anywhere in the check, we only use it when
remediating the rule.

To work around this we define an external variable just to set
the type of the variable to be as SCAP specification defines.