Fix installation path of guides and roles
When doing an install where CMAKE_INSTALL_DOCDIR is specified, the
install directory for roles and guides become like this
"...scap-security-guide-0.1.33-1.el7.noarch/usr//usr/share/doc/..."

GNUInstallDirs.cmake checks wether CMAKE_INSTALL_DOCDIR is absolute
and add prefix CMAKE_INSTALL_PREFIX if not.
But it is not applied when CMAKE install uses CODE.

So we check if SSG_ROLE_INSTALL_DIR and SSG_GUIDE_INSTALL_DIR is
absolute or not and prefix the CMAKE_INSTALL_PREFIX if necessary.