SubElement would cause 2 appends which is not what we want
We previously used libxml2 implementation of the ElementTree API which
didn't permit one instance of the same element appearing multiple times
in the parent. The python2 inbuilt ET allows this, so we can no longer
double append. This was all done by mistake so it's not a big deal to
just change SubElement to Element.