Make schematron OVAL validation optional but default it to true
Schematron takes forever, it is by far the most time consuming part of
the jenkins run. As we try to make CI better for everyone in the
community we want to get under 5 mins for a jenkins run. Our plan is to
run schematron validations nightly but not use them for PR gating.
Schematron validation of a single big OVAL file takes ~4 minutes on very
good hardware.