Workaroung handling of fixparts in pairs
combine-remediations.py handle fixparts in pairs.
If two function calls are put after each other, the second call of the
function will not be expanded, just appended as pure text.

By adding newlines we workaround this issue, the newline is the
second element in the pair, and is just fine being appended.