Also check state openssh-server package when sshd_required is unset
Explicitly check state of openssh-server package.
When openssh-server is installed, system should be configured, when not
installed, system is ok.
When sshd_required is set, either to required or not required, they act
as selector of openssh-server package state. If sshd_required is unset,
the state of openssh-server package selects whether system should be
configured or not.