Do not fail aide_scan_notification with other email adresses (#2599)
* Do not fail aide_scan_notification with other email adresses

The rule aide_scan_notification says that AIDE should notify appropriate
personnell of the detials of an AIDE scan. The check currently requires
that the email address of the appropriate personell starts with 'root@'.
In practice, the email address could be any email address. The check
should match any email address.
Fixes RHBZ#1540505