Improvments for banner checks and remediations (/etc/issue and GDM)
This commit optimizes the regular expressions for the different available
banners and also the sed transformations used in order to produce the
final configuration files for /etc/issue and GDM banner.

The transformations are written in order to handle the specific cases
which are transforming \n and ' characters. A specific regular
expression has been writeen for each case (/etc/issue and GDM)