Fix append scenario in remediation for audit_data_retention_space_left
Remediation for auditd_data_retention_space_left could not add
configuration line when it was missing.

grep command would return 0, sed command would not find the
configuration line to replace and return 0.
Subsequent if would consume $? equal 0 and not append config.

Use of function replace_or_append is not easy here, [[:space:]] is not
handled correctly.