When enabling FIPS, set boot= arg using /boot UUID.
When enabling FIPS mode, the content assumes vmlinuz
images are in /boot/efi. They appear to remain in /boot.
Additionally, relying on device node paths is fragile.

Grab the UUID of /boot and use that to populate the
kernel command-line when enabling FIPS mode.

Fixup a few of the regexs to make them easier to
comprehend.

Only disable prelinking only if the RPM is present.