Add linkcheckerrc to set sslverify=0
This prevents us from validating the certificates of the links we check.
Because https://pcisecuritystandards.org does not serve the intermediate
certificate chain, we won't pass linkchecker tests with sslverify=1 (the
default). On the other hand, most browsers will cache intermediate CA
certs, letting the user access this website. Maintaining our own CA chain
just to add the GoDaddy intermediate cert we need is both fragile and
too much work for the little value it provides, so this is a better
compromise.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>