Use implied property to specify dnf/yum config file
Some rules describe package manager configuration. However, on RHEL we
use yum and we set configuration settings in /etc/yum.conf, but on
Fedora we use dnfand we set configuration in /etc/dnf/dnf.conf.
We can use implied properties to customize the rules and remediations.
These implied properties will be substituted in final output, so there
will be "dnf" mentioned in Fedora content and "yum" mentioned in RHEL
content. Also, the rules and remediations will contain correct path
to the configuration file depending on the package manager used
on target platform.