Remove the superfluous OS checks from OVAL
The RHEL 6 part of this OVAL that checked pam_cracklib instead of
pam_pwquality has been splitted out to a separate rule
"password_quality_pamcracklib". In this rule we check for
pam_pwquality settings only. It is not needed to check for operating
systems installed anymore. It does not change the results and it
complicates the adoption of this rule in other products. The profile
authors should choose a correct rule in their profile selection.