Changed shell command
Changed the shell command per @cmattern-rht .  Not all values returned with <code>rpm -Va | grep ^.M </code> have the same number of fields and so the <code>rpm --setperms</code> was failing on those empty values. This will take the last value which is always the path that is needed.