New rule and group field platform
It allows us to mark rules as bare-metal machine only or container
only by adding "platform: machine" or "platform: container" to
rule.yml or group.yml. This will be populated into XCCDF as platform
element. The platform element is a child element of Rule or Group
element. For convenience the CPE in idref attribute value is defined
as a constant (under ssg/constants.py).