Fix SSG Test Suite for Pythoon 3.7
We use `exclude` argument of `Tarfile.add()` method. This argument
is deprecated since Python 2.7 and Python 3.2 and has been removed
in Python 3.7. See https://bugs.python.org/issue29210. Instead, we
should use `filter` argument which is available in Python 2.7, 3.2
and newer. This fixes SSG test suite on Python 3.7 which is used
in Fedora 29.