Introduce platform dependent test scenarios
We face a problem that for a single rule we need to have different sets
of test scenarios depending on target platform. For example, the rule
'auditd_audispd_encrypt_sent_records' exists on both RHEL 7 and RHEL 8,
but on each of them a different file path and a different string
is checked. We need to be able to define test scenarios that are
applicable only to selected platforms and should be skipped when testing
other platforms.

This will be achieved by using platform option in scenario script headers.
It works similar way we specify platforms in Bash remediations. We can
use `multi_platform_rhel`, `multi_platform_fedora`, etc.