Enhance configure_kerberos_crypto_policy check
The goal of this check is to verify that /etc/krb5.conf.d/crypto-policies is
a symlink, and points to /etc/crypto-policies/back-ends/krb5.config.

As the symlink_test goes all the way through to the canonical path,
and the canonical path is the actual selected policy configuration, thus check
was dependent on selected policy and was failing when symlink was
correct, but selected crypto policy was wrong.

Making sure that /etc/krb5.conf.d/crypto-polices links to correct crypto-policy
is not the poinof this check. With this changes, the check is now
verifying the symlink independently of the selected crypto policy.