Remove kubelet_enable_authorization
Unauthenticated/Unauthorized users have no access to OpenShift nodes.  The kubelet flag authorization-mode is explicitly set to WebHook

Since Webhook is the default value, the authorization-mode configuration is not needed in node-config.  If it is configured, ensure the value is only set to ‘Webhook’.  The node systemd service (atomic-openshift-node) will not start if the configuration value is set to anything other than Webhook.