Remove test scenario missing-system-auth.fail.sh
This scenario is completely invalid. We have some rules, where missing
config file is finding that is fixable. Usually it's to override some
default values. PAM configurations are not for overriding defaults, but
to define behaviour. So in case system-auth is missing, either machine
is completely broken and in very unsupported configuration, or not using
PAM. We shouldn't try to fix it in such case.