Build Ansible Playbooks for each rule in every profiles
For each profile it creates Ansible Playbooks for each rule that has an
Ansible remediation available.  The build is a part of product target
for all products.  The Playbook contains values of XCCDF Values selected
in the profile.
The output is generated to `/build/<product>/playbooks/<profile>`.
Each Playbook is in a separate file `<rule_id>.yml`.