Fix remediation scripts terminology
Currently, we build Ansible Playbooks for profiles into `roles`
directory and they're called `ssg-product-role-profile.yml`. That can
make a false impression that those files are Ansible roles, but they are
not Ansible roles, they are Ansible Playbooks.  Similarly, we now build
"Bash roles". But the term "Bash role" doesn't exist, the Bash files
that we build should be called "Bash remediation scripts" instead. This
patch fixes this confusion of terms.  The Ansible Playbooks will be
built into `playbooks` directory and they will have `playbook` instead
of `role` in their name. The Bash remediation scripts will be built into
`bash-scripts` directory and they will have `script` instead of `role`
in their name. Hopefully it will be less confusing for users.
The playbook and script headers are also confusing but that has already
been fixed in OpenSCAP in
https://github.com/OpenSCAP/openscap/pull/1261.