Remove documentation_complete attribute as it is not handled by the profile.
Make the profile subtraction simpler.

The documentation_complete attribute only makes sense when dumping
Profile to yaml and during processing of yaml profile, this attribute
is popped from input data and is not possible to keep it there in a
simple way. So the user when dumping the yaml can decide whether he
wants to set documentation_complete as true or false. Default is true.