install_vm.py: leave default disk location to libvirtd
Libvirt uses "storage pools" to automatically allocate storage for VMs.
These may be files on disk, LVM LVs, glusterfs volumes and others.

Therefore leave the default as size=20 only as we don't care how the
VM image will be stored.

This should also automagically work with non-root users as libvirt
will know to store non-root images in the user homedir.

The format is left hardcoded as we need qcow2 for internal snapshots
for test_suite.py.

Signed-off-by: Jiri Jaburek <jjaburek@redhat.com>