Remove the 'build' directory by default
Most OSS projects that use cmake let the user or a script create
the directory instead of it being pre-created.

While it might seems just like a convenience here, if you actually
remove the directory (to make a new build) and later commit with
`git commit -a', the removal of `build/.gitkeep' will be added to
the commit.

.gitignore does not help here as it only prevents untracked files
from being added - changes in existing tracked files are still
registered.

While there are other local-only workarounds like

  git update-index --skip-worktree build/.gitkeep

a cleaner solution is removing the default build dir altogether.

Signed-off-by: Jiri Jaburek <jjaburek@redhat.com>