Quietly handle documentation not complete errors
When the documentation isn't complete, ssg.yaml would silently return
None. This means that code would just detect if the result was None, not
knowing why (it could be an empty profile).

Instead, we raise a DocumentationNotComplete error: this will tell
callers why the result was None and what to do to fix it (run in debug
mode). In the most common output, this now shows:

    Skipping incomplete profile rhel7/profiles/anssi_nt28_minimal.profile.
    To include incomplete profiles, build in debug mode.

Instead of the former message:

    Could not parse profile rhel7/profiles/anssi_nt28_intermediary.profile.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>