Override _parse_parameter() to override metadata
In CombinedChecker, all target rules are already part of a profile.
Let's override 'profiles' metadata in test scenarios so they can be
tested cleanly.

Note: This automagically will enable all test scenarios of rules to the
target profile. There can be test fails if the test scenarios are
setting up values not compliant with the target profile.

This can be solved by improving the test scenarios to setup a range of
values that always will be pass/fail as expected. Or with some Jinja2
macros.

I believe the benefits of this change outweight the disadvantages.