Change default separator regex
The former default separator regex `(?-i)[\s]+` was wrong, because
`(?-i)` means "stop being case-insensitive". It needs a pair `(?i)`
before in the regular expression and is ineffective but confusing alone.
It should be harmless to change the default value now because all of the
occurrences are calling the macro with the parameter set.