Compare suid/sgid files with the RPM database
It is difficult to maintain the list to list paths of all possible suid
and sgid binaries in a Linux distribution. Instead, we can check if the
suid or sgid file is owned by an RPM package by consulting the RPM
database. Another advantage of this solution is that we can have a
single OVAL for all RPM-related Linux distributions.  The patch modifies
OVAL for rules file_permissions_unauthorized_suid and
file_permissions_unauthorized_sgid and also adds test scenarios for
these rules.
Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=1693026