Pass target SCAP version to update script
Enpower DS version update script to decide when to update or not,
instead of controling this decision on CMake script

This is to avoid defining 4 custom commands to build the DS
in CMake macro ssg_build_ds(). Namely:
1. command for rhel6|7 with 1.2 (includes pci-dss benchmark)
2. command for rhel6|7 with 1.3 (includes pci-dss benchmark)
3. command for other products with 1.2
4. command for other produts with 1.3

The script receives target SCAP version and handles update of version.
When target version is 1.3, it is updated.
When target version is 1.2, nothing is done.
Also, renamed script to generic update sds version.