Fix STIG IDs reference processing.
This PR allows to assign a rule multiple STIG IDs using a
comma-separated list which is the same way we use for the other
references. There were 2 basic problems: 1. Python code converting YAML
to shorthand was not splitting the string at comma. 2. XSLT template
that converts shorhand to XCCDF had a special treatment for STIG IDs
which always copies the whole attribute value instead of each identifier
which lead to duplicate `reference` elements.