Prevent failures of Ansible task for service_enabled
If the package which provides the service is not present the Ansible
task that enables the service will fail because it can not find the unit
file which will cause termination of the whole playbook. We can prevent
the failure by executing the task only if the given package is present.
The code uses Ansible package module in check mode to determine
if package is present.
Partially fixes: https://bugzilla.redhat.com/show_bug.cgi?id=1741455
Addressing:
TASK [Enable service fapolicyd]
************************************************
fatal: [192.168.122.24]: FAILED! => {"changed": false, "msg": "Could not
find the requested service fapolicyd: host"}