Excluding /run and /sys* from filesystems watched by the fapolicyd.
We don't want to create too many events when we don't care about
some filesystems. The more we watch, the slower the system gets.
Nothing in /sys that is writable by a normal user, it's for sysctls.
The /run directory is owned by root and all subdirs have restricted
permissions.