Merge all csv_data into shared and resolve it
When a product's csv_data is merged into another product's csv_data,
every different value for a parameter is added into a product counter.
This counter keeps track of how many and which products use the value.

When resolving the csv_data, these counters are used to determine the
shared value, and the product specific values for a parameter.