Skip content generation when callback isn't implemented
Currently, if the callback is not present the data are passed directly
to Jinja which sometimes lead to generating invalid OVAL and build
cannot be finished. This change will enable us to add the callbacks one
by one in multiple separate pull requests and we will be able to
complete the build using the new templating system without errors.