Fix disappeared OVALs for audit_rules_execution.* rules
Actually in the old system the same template
shared/templates/template_OVAL_audit_rules_privileged_commands  was
shared between 2 different create_*.py scripts
shared/templates/create_audit_rules_execution.py and
shared/templates/create_audit_rules_privileged_commands.py.
The 2 scripts differed in populating the ID variable which was used as
the OVAL definition ID in the template. Since the OVAL definition ID
is expected to be same as the rule ID, we can add the rule ID to the
Jinja variables and use it in the template. Later, we should do that
in all other templates, because we don't want to rely on this implicit
rule ID matching.