Add some variables to the env_yaml dict
When we process the Jinja macros in OVALs or remediations which are
defined in rule directories we add variables (eg. rule title, platforms)
to the Jinja environment. The Jinja macros that are used in templates
for auditd_lineinfile and templates for sshd_lineinfile require that
these variables are defined. If they aren't defined they silently
generate an invalid OVAL which is then dropped later in the build
process.  Therefore we need to define these variables also in the
env_yaml that is used in the templating engine to get a similar
environment for generating content from templates as when processing
not templated content.  This is similar to code in checks() in
ssg/build_ovals.py.