Generate OVAL definition without need for a rule
This patch introduces a way to generate templated OVALs
that aren't "owned" by any rule.

Some templated OVALs are needed as dependency in other OVALs
referenced by extend_definition. The new templating system
is based on rules, so to generate an OVAL definition a rule
should be created. But for some OVALs it isn't acceptable to
have a special rule in the benchmark.

This is a temporary solution until applicability of all rules
is handled by using CPE applicability checks.

See https://github.com/ComplianceAsCode/content/pull/4874