Remove overly restrictive FIPS checks
Whether or not a package is configured to use FIPS approved crypto
algorithms and hashes or fips-related packages are installed is orthogonal
to whether or not the installed operating system is FIPS certified.  This
removes the OS certification test from the configuration checks for
individual components.  Whether or not the OS is FIPS certified still
remains a stand alone rule by itself and should continue to be used in
environments where necessary.