Fix platform in OVAL for accounts_logon_fail_delay
This rule is used in OCP4, RHEL7, OL7, RHV4, WRLinux 10.19 and WRLinux 8
profiles. However, the OVAL was available only for RHEL 7 and WRLinux 8
which is not aligned with empty element in rule.yml. Since the rule
makes sense on most of the operating systems, we can mark the OVAL as
multi_platform_all.

The rhel7.xml and wrlinux.xml were almost identical, except the platform
element, so we can merge them in a single shared.xml.