Remove openshift api_server_profiling check from ocp3 CIS profile
The ocp3 opencis-master profile has a check that expects profiling
for the API server to be explicitly disabled.  Red Hat has changed
their guidance in this area, as API server profiling is protected
via RBAC and can be left enabled or disabled as desired.  From a
security perspective, we only want to make sure that the web
interface for profiling is disabled for the controller.  We already
have the controller_disable_profiling check for this.

This patch removes the api_server_profiling check from the ocp3
CIS profile.