Update test scenarios for chronyd_or_ntpd_set_maxpoll for RHEL8
The rule chronyd_or_ntpd_set_maxpoll checks both ntp and chronyd
configuration. The test scenarios have been developed on RHEL7, but the
rule is also applicable to RHEL8. But on RHEL8, there is no ntp, only
chronyd. Therefore the scenarios that test ntp configuration should not
be applicable to RHEL8.

Also, the rule checks `maxpoll` parameters only on `server` keys, but
`server` keys are not the only way to configure time source, there can
be also `pool` and other keys. On RHEL8, the default `/etc/chrony.conf`
does not contain `server` but it contains `pool`.  That means to test
the `server` key we cannot rely on its default presence but we need to
insert a line if it is not present.