ocp4/profiles/coreos-ncp.profile: disable all the rules
The coreos-ncp profile was suggested to use for FedRAMP moderate
certification of OCP4.

But many of the rules that the profile reference are not applicable to
the OCP4 product and quite a few of them even should not be (e.g. -
package_sssd-ipa_installed will most certainly not be useful for the
general installation of OCP4).

We can either remove or comment out all the rules. I opted to comment
them out as it might be useful to reference them later when checking if
an individual rule should apply for the RHCOS host or not.